+++ src/main/java/com/appland/appmap/process/hooks/ToggleRecord.java
@@ -193,6 +193,16 @@ public class ToggleRecord {
   }
 
   @ArgumentArray
+  @CallbackOn(MethodEvent.METHOD_EXCEPTION)
+  @ExcludeReceiver
+  @HookAnnotated("org.junit.Test")
+  public static void junit(Event event, Exception exception, Object[] args) {
+    event.setException(exception);
+    recorder.add(event);
+    stopTest();
+  }
+
+  @ArgumentArray
   @ExcludeReceiver
   @HookAnnotated("org.junit.jupiter.api.Test")
   public static void junitJupiter(Event event, Object[] args) {
@@ -204,6 +214,16 @@ public class ToggleRecord {
   @ExcludeReceiver
   @HookAnnotated("org.junit.jupiter.api.Test")
   public static void junitJupiter(Event event, Object returnValue, Object[] args) {
+    stopTest();
+  }
+
+  @ArgumentArray
+  @CallbackOn(MethodEvent.METHOD_EXCEPTION)
+  @ExcludeReceiver
+  @HookAnnotated("org.junit.jupiter.api.Test")
+  public static void junitJupiter(Event event, Exception exception, Object[] args) {
+    event.setException(exception);
+    recorder.add(event);
     stopTest();
   }
 }