+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2020-07-30
+### Fixed
+- Added callback handlers with `METHOD_EXCEPTION` method event to close recording of current session
+ for `Junit` and `Jupiter` Tests.
+  
 ## [0.3.1] - 2020-07-15
 ### Fixed
 - Dependencies of `appmap-java` which use `slf4j` will now use a nop implementation in order